examples/sqlite/pyproject.toml
[project]
name = "app"
description = ""
keywords = []
dynamic = [ "version" ]

requires-python = ">=3.10,<3.14"

dependencies = [
    "dynaconf-tortoise-loader>=0.1.0"

]
[tool.setuptools.dynamic]
version = {attr = "app.__version__"}

[tool.setuptools.packages.find]
where = ["src"]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"
